cloud_core/requirements.txt
# Core requirements for cloud testing
pandas>=2.3.2
numpy<2.0.0
scikit-learn>=1.4.0
xgboost>=2.0.0
hmmlearn>=0.3.0
pandas-ta
ccxt>=4.3.0
httpx>=0.27.0
python-telegram-bot>=20.8
joblib
python-dotenv>=1.0.0
lighter-sdk>=1.0.0
eth_account>=0.11.0
aiohttp>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
